+++ application/models/Vendedor.php
@@ -215,7 +215,7 @@ class Vendedor extends CI_Model{
 
         $vendedor = $this->getVendedorPorCodigo($codVendedor);
 
-        $input = number_format($vendedor->perc_comissao, 2, ',', '.');
+        $input = number_format((float) ($vendedor->perc_comissao), 2, ',', '.');
 
         return $input;
